db/mysql/scripts/mysql_manager/requirements.txt
mysql-connector-python==8.4.0
packaging==24.1
python-dotenv==1.0.1
